requirements.txt
Django
djangorestframework
coreapi
django-filter
geopy